requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
amqp==5.3.1
arrow==1.3.0
asgiref==3.8.1
attrs==24.3.0
billiard==4.2.1
binaryornot==0.4.4
black==24.10.0
Brotli==1.1.0
celery==5.4.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorlog==6.9.0
cookiecutter==2.6.0
Django==5.1.4
django-jazzmin==3.0.1
django-modeltranslation==0.19.11
django-unfold==0.43.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
frozenlist==1.5.0
g4f==0.3.9.7
idna==3.10
inflection==0.5.1
isort==5.13.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jst-aicommit==1.3
jst-django==4.0.6
jst-django-core==1.0.9
kombu==5.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
packaging==24.1
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
polib==1.2.0
prompt-toolkit==3.0.36
propcache==0.2.1
pycryptodome==3.21.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.2
questionary==2.0.1
referencing==0.35.1
requests==2.32.3
rich==13.9.3
rpds-py==0.22.3
shellingham==1.5.4
six==1.16.0
sqlparse==0.5.3
tenacity==9.0.0
text-unidecode==1.3
tqdm==4.67.1
typer==0.12.5
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.3
zipfile36==0.1.3
